cleanup linting part 5
<!-- ps-id: a74cad4d-03e9-47f5-97ac-056e42031e59 -->

+++ compiler/emitter.go
@@ -215,7 +215,7 @@ func (e *Emitter) FixupJumps() error {
 			}
 		case OperandRelative32:
 			// Safe conversion with explicit truncation
-			inst.Operand.Value = uint64(relativeOffset)
+			inst.Operand.Value = uint64(int32(relativeOffset)) // #nosec G115
 		default:
 			return fmt.Errorf("unsupported operand type for jump fixup: %v", inst.Operand.Type)
 		}
